+++ src/components/TaskModal/styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  pressableOverlayArea: {
+    backgroundColor: 'rgba(0,0,0,.4)',
+    flex: 1,
+  },
+  textInput: {
+    backgroundColor: 'white',
+    borderRadius: 10,
+    margin: 10,
+    padding: 10
+  },
+});
